Update prob 1360B, 1360C - Codeforces

+++ Codeforces/1360C.cpp
@@ -10,13 +10,14 @@ typedef long long ll;
 int main(){
     ios_base::sync_with_stdio(false);
     cin.tie(0);
+
     ll t;
     cin >> t;
     while(t--) {
         ll n; cin >> n;
         vector<ll> arr;
         ll num_even = 0, num_odd = 0, distance = 0;
-        for(int i = 0;i < n;++i){
+        for(ll i = 0;i < n;++i){
             ll value; cin >> value;
             arr.push_back(value);
             if(value % 2 == 0){
@@ -30,8 +31,8 @@ int main(){
             cout << "YES" << endl;
         }
         else{
-            unordered_multiset<int> distance(arr.begin(), arr.end());
-            for(int i = 0;i < n;i++){
+            unordered_multiset<ll> distance(arr.begin(), arr.end());
+            for(ll i = 0;i < n;i++){
                 if(distance.find(arr[i]) != distance.end() && distance.find(arr[i] + 1) != distance.end()){
                     --num_odd; --num_even;
                     break;
